backend_python/requirements.txt
Flask~=2.3.3
pymongo~=4.10.1
Flask-JWT-Extended==4.5.1
Flask-SocketIO==5.3.3
eventlet==0.33.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
gunicorn==20.1.0
python-dotenv==1.0.0
requests~=2.32.3
mongoengine==0.29.1
PyJWT~=2.10.1
bcrypt~=4.2.1
dnspython~=2.7.0
idna~=3.10
h11~=0.14.0
pip~=24.2
protobuf~=5.29.3
filelock~=3.16.1
typing_extensions~=4.12.2
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
Jinja2~=3.1.6
click~=8.1.8
blinker~=1.9.0
itsdangerous~=2.2.0
certifi~=2025.1.31
urllib3~=2.4.0
python-engineio~=4.12.0
tornado~=6.4.2
simple-websocket~=1.1.0
six~=1.17.0
greenlet~=3.2.0
charset-normalizer~=3.4.1
bidict~=0.23.1
python-socketio~=5.13.0
setuptools~=75.8.0
more-itertools~=10.6.0
wsproto~=1.2.0